requirements.txt
django==4.2.5
python-dotenv==1.0.0
psycopg2-binary==2.9.7
pyyaml==6.0.1
djangorestframework==3.14
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
django-filter==23.3
drf-yasg==1.21.7

